create a new survey with loader class for when viewing survey by itself
